Differentiate between viewed and unviewed notices
This appends an asterisk to the notification count if one has viewed the
notification(s) i.e., clicked on the GH-N in the mode line, since they
were received.

When a new notification is received the asterisk is removed.

This allows one to differentiate between notifications that they've viewed
and/or read and new, unviewed, ones.